+++ Advanced C++ Graphics T Pt 2/Zombie.cpp
@@ -21,25 +21,19 @@ Zombie::~Zombie()
 }
 
 
-
-/* NEW: added stance arguments */
 void Zombie::init(float speed, glm::vec2 pos, const std::unordered_map<unsigned int, GLuint>& stancesIDs, const NumStances& numStances) {
 	m_speed = speed;
 	m_position = pos;
 	m_health = 150;
 
-	/* NEW: change color so solid white */
-	//m_color = Bengine::ColorRGBA8(0, 160, 0, 255);
+
 	m_color = Bengine::ColorRGBA8(255, 255, 255, 255);
 
 
 
-
-	/* NEW */
 	m_numStances = numStances;
 	m_stanceIDs = stancesIDs;
 	m_textureID = m_stanceIDs[m_currentMoveImage];
-	/* NEW: end of new */
 
 
 }
@@ -51,15 +45,7 @@ void Zombie::update(const std::vector<std::string>& levelData,
 	std::vector<Zombie*>& zombies,
 	float deltaTime) {
 
-
-
-	/* NEW */
 	spriteStanceUpdate();
-
-
-
-
-
 
 
 	static std::mt19937 randomEngine(time(nullptr));
@@ -77,16 +63,8 @@ void Zombie::update(const std::vector<std::string>& levelData,
 	Human* closestHuman = getNearestHuman(humans);
 
 	if (closestHuman != nullptr) {
-
-		/* NEW: now agent protected variable */
-		//glm::vec2 direction = glm::normalize(closestHuman->getPosition() - m_position);
 		m_direction = glm::normalize(closestHuman->getPosition() - m_position);
-
-		/* NEW */
-		//m_position += direction * m_speed * deltaTime;
 		m_position += m_direction * m_speed * deltaTime;
-
-
 	}
 	else {
 		//TO DO: implement
@@ -113,7 +91,7 @@ Human* Zombie::getNearestHuman(std::vector<Human*>& humans) {
 	float smallestDistance = 9999999.0f;
 
 
-	for (unsigned int i = 0; i < humans.size(); ++i) {
+	for (size_t i = 0; i < humans.size(); ++i) {
 		glm::vec2 distVec = humans[i]->getPosition() - m_position;
 		float distance = glm::length(distVec);
 
@@ -132,15 +110,12 @@ Human* Zombie::getNearestHuman(std::vector<Human*>& humans) {
 
 
 
-
-/* NEW */
 void Zombie::spriteStanceUpdate(Uint32 currentSpriteStance /*  = SpriteStance::MOVING */) {
-	/* NEW */
+
 	++m_currentMoveImage;
-	/* NEW: moving image */
+
 	if (m_currentMoveImage > m_numStances.lMove) {
 		m_currentMoveImage = m_numStances.fMove;
 	}
 	m_textureID = m_stanceIDs[m_currentMoveImage];
-	/* NEW: end of new */
 }
