+++ Bengine/SpriteBatch.cpp
@@ -7,7 +7,6 @@
 
 namespace Bengine {
 
-	/* NEW: overloaded constructor, for background sprites like blood particles */
 	Glyph::Glyph(const glm::vec4& destRect, const glm::vec4& uvRect, GLuint Texture, float Depth, const ColorRGBA8& color) :
 		texture(Texture),
 		depth(Depth) {
@@ -29,70 +28,50 @@ namespace Bengine {
 
 
 
-	/* NEW: overloaded constructor */
+
 	Glyph::Glyph(const glm::vec4& destRect, const glm::vec4& uvRect, GLuint Texture, float Depth, const ColorRGBA8& color, float angle) :
 		texture(Texture),
 		depth(Depth) {
 
-		/* NEW: orgin, usually top left is treated as position */
+
 		glm::vec2 halfDims(destRect.z / 2.0f, destRect.w / 2.0f); 
-		/* NEW: vec3 is for 3D games, get points centered at orgin */
 		glm::vec2 tl(-halfDims.x, halfDims.y); // tl = topleft
 		glm::vec2 bl(-halfDims.x, -halfDims.y); // bl = bottomleft
 		glm::vec2 br(halfDims.x, -halfDims.y); // br = bottomright
 		glm::vec2 tr(halfDims.x, halfDims.y); // tr = topright
-		/* NEW: rotate the points, add back half dims to treat top left as position and not orgin */
+
 		tl = rotatePoint(tl, angle) + halfDims;
 		bl = rotatePoint(bl, angle) + halfDims;
 		br = rotatePoint(br, angle) + halfDims;
 		tr = rotatePoint(tr, angle) + halfDims;
 
 
-
-
 		// z represents width and w represents height
 		topLeft.color = color;
-
-		/* NEW */
-		//topLeft.setPosition(destRect.x, destRect.y + destRect.w);
 		topLeft.setPosition(destRect.x + tl.x, destRect.y + tl.y);
-
 		topLeft.setUV(uvRect.x, uvRect.y + uvRect.w);
 		bottomLeft.color = color;
-
-		/* NEW */
-		//bottomLeft.setPosition(destRect.x, destRect.y);
 		bottomLeft.setPosition(destRect.x + bl.x, destRect.y + bl.y);
-
 		bottomLeft.setUV(uvRect.x, uvRect.y);
 		bottomRight.color = color;
-
-		/* NEW */
-		//bottomRight.setPosition(destRect.x + destRect.z, destRect.y);
 		bottomRight.setPosition(destRect.x + br.x, destRect.y + br.y);
-
 		bottomRight.setUV(uvRect.x + uvRect.z, uvRect.y);
 		topRight.color = color;
-
-		/* NEW */
-		//topRight.setPosition(destRect.x + destRect.z, destRect.y + destRect.w);
 		topRight.setPosition(destRect.x + tr.x, destRect.y + tr.y);
-
 		topRight.setUV(uvRect.x + uvRect.z, uvRect.y + uvRect.w);
 	}
 
 
 
 
-	/* NEW */
+
 	glm::vec2 Glyph::rotatePoint(glm::vec2 pos, float angle) {
 
-		/* NEW */
 		glm::vec2 newv; // new vector
 		newv.x = pos.x * cos(angle) - pos.y * sin(angle);
 		newv.y = pos.x * sin(angle) + pos.y * cos(angle);
 		return newv;
-		/* NEW: end of new */
+
 	}
 
 
@@ -133,7 +112,7 @@ namespace Bengine {
 
 	void SpriteBatch::end() {
 		m_glyphPointers.resize(m_glyphs.size());
-		for (unsigned int i = 0; i < m_glyphs.size(); ++i) {
+		for (size_t i = 0; i < m_glyphs.size(); ++i) {
 			m_glyphPointers[i] = &m_glyphs[i];
 		}
 		sortGlyphs();
@@ -151,7 +130,6 @@ namespace Bengine {
 
 
 
-	/* NEW: overloaded method */
 	void SpriteBatch::draw(const glm::vec4& destRect, const glm::vec4& uvRect, GLuint texture, float depth, const ColorRGBA8& color, float angle) {
 		m_glyphs.emplace_back(destRect, uvRect, texture, depth, color, angle);
 	}
@@ -159,7 +137,7 @@ namespace Bengine {
 
 
 
-	/* NEW: overloaded method */
+
 	void SpriteBatch::draw(const glm::vec4& destRect, const glm::vec4& uvRect, GLuint texture, float depth, const ColorRGBA8& color, const glm::vec2& dir) {
 
 		const glm::vec2 right(1.0f, 0.0f);
@@ -177,7 +155,7 @@ namespace Bengine {
 
 	void SpriteBatch::renderBatch() {
 		glBindVertexArray(m_vaoID);
-		for (unsigned int i = 0; i < m_renderBatches.size(); ++i) {
+		for (size_t i = 0; i < m_renderBatches.size(); ++i) {
 			glBindTexture(GL_TEXTURE_2D, m_renderBatches[i].texture);
 			glDrawArrays(GL_TRIANGLES, m_renderBatches[i].offset, m_renderBatches[i].numVertices);
 		}
@@ -212,7 +190,7 @@ namespace Bengine {
 
 		offset += 6;
 
-		for (unsigned int cg = 1; cg < m_glyphs.size(); ++cg) { //current Glyph
+		for (size_t cg = 1; cg < m_glyphs.size(); ++cg) { //current Glyph
 			if (m_glyphPointers[cg]->texture != m_glyphPointers[cg - 1]->texture) {
 				m_renderBatches.emplace_back(offset, 6, m_glyphPointers[cg]->texture);
 			}
