+++ Bengine/SpriteFont.cpp
@@ -92,7 +92,7 @@ namespace Bengine {
 		int ly = padding;
 		for (int ri = 0; ri < bestRows; ri++) {
 			int lx = padding;
-			for (unsigned int ci = 0; ci < bestPartition[ri].size(); ci++) {
+			for (size_t ci = 0; ci < bestPartition[ri].size(); ci++) {
 				int gi = bestPartition[ri][ci];
 
 				SDL_Surface* glyphSurface = TTF_RenderGlyph_Blended(f, (char)(cs + gi), fg);
