+++ Bengine/Timing.cpp
@@ -1,6 +1,7 @@
 #include "Timing.h"
 
-#include <SDL/SDL.h>
+/* NEW: moved to Timing.h */
+//#include <SDL/SDL.h>
 
 
 
@@ -46,8 +47,14 @@ namespace Bengine {
 
 	float FpsLimiter::end() {
 		calculateFPS();
+
 		// Limits the frame rate
-		float frameTicks = (float)SDL_GetTicks() - (float)m_startTicks;
+
+		/* NEW: changed to Uint32 */
+		//float frameTicks = (float)SDL_GetTicks() - (float)m_startTicks;
+		Uint32 frameTicks = SDL_GetTicks() - m_startTicks;
+
+
 		if (1000.0f / m_maxFPS > frameTicks) {
 			SDL_Delay((Uint32)(1000.0f / m_maxFPS - frameTicks));
 		}
@@ -63,12 +70,30 @@ namespace Bengine {
 
 	void FpsLimiter::calculateFPS() {
 		static const int NUM_SAMPLES = 10;
-		static float frameTimes[NUM_SAMPLES];
+
+
+		/* NEW: changed to Uint32 */
+		//static float frameTimes[NUM_SAMPLES];
+		static Uint32 frameTimes[NUM_SAMPLES];
+
+
 		static int currentFrame = 0;
 
-		static float prevTicks = (float)SDL_GetTicks();
-		float currentTicks;
-		currentTicks = (float)SDL_GetTicks();
+
+		/* NEW: changed to Uint32 */
+		//static float prevTicks = (float)SDL_GetTicks();
+		static Uint32 prevTicks = SDL_GetTicks();
+
+
+		/* NEW: changed to Uint32 */
+		//float currentTicks;
+		Uint32 currentTicks;
+
+		/* NEW: changed to Uint32 */
+		//currentTicks = (float)SDL_GetTicks();
+		currentTicks = SDL_GetTicks();
+
+
 		m_frameTime = currentTicks - prevTicks;
